fix(UI-1800): hubspot tracking
