+++ src/hooks/index.ts
@@ -17,7 +17,6 @@ export { useTourActionListener } from "./tour/useTourActionListener";
 export { useEvent } from "./useEvent";
 export { useRateLimitHandler } from "./useRateLimitHandler";
 export { useChatbotIframeConnection } from "./useChatbotIframe";
-export { useHubspotSubmission } from "./useHubspotSubmission";
 export { useLoginAttempt } from "./useLoginAttempt";
 export { useBilling } from "./billing/useBilling";
 export { useProjectMetadataHandler } from "./useProjectActionsResponseHandler";
