+++ src/interfaces/hooks/index.ts
@@ -1,5 +1,4 @@
 export type { ResizeHook } from "@interfaces/hooks/useResize.interface";
 export type { VirtualizedListHookResult } from "@interfaces/hooks/useVirtualizedList.interface.ts";
 export type { IframeState, IframeError } from "@interfaces/hooks/useChatbotHook.interface";
-export type { HubspotSubmissionArgs } from "./useHubspotSubmission.interface";
 export type { UseLoginAttemptArgs } from "./useLoginAttempt.interface";
