--- src/interfaces/hooks/useHubspotSubmission.interface.ts
@@ -1,5 +1,0 @@
-import type { TFunction } from "i18next";
-
-export interface HubspotSubmissionArgs {
-	t: TFunction;
-}
